backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mediapipe==0.10.7
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0

